@@ -1,3 +1,3 @@
-module github.com/uber/h3-go/v4
+module github.com/bulatsan/h3-go/v4
 
-go 1.18
+go 1.21
